Refactor: Move files to functions register
Added trigger functions
