+++ Ofqual.Common.RegisterAPI.sln
@@ -3,9 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.9.34607.119
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "QualificationsRegister", "QualificationsRegister\QualificationsRegister.csproj", "{6A6BE76A-18F0-44EA-A672-8B1124BF28D1}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "QualificationsRegister", "QualificationsRegister\QualificationsRegister.csproj", "{6A6BE76A-18F0-44EA-A672-8B1124BF28D1}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "QualificationsRegister.Tests", "QualificationsRegister.Tests\QualificationsRegister.Tests.csproj", "{ED60365D-9B05-4FBD-8ED3-633048696A7F}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "QualificationsRegister.Tests", "QualificationsRegister.Tests\QualificationsRegister.Tests.csproj", "{ED60365D-9B05-4FBD-8ED3-633048696A7F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
