+++ QualificationsRegister/Functions/QualificationsInternal.cs
@@ -6,13 +6,13 @@ using Ofqual.Common.RegisterAPI.UseCase.Interfaces;
 
 namespace Ofqual.Common.RegisterAPI.Functions
 {
-    public class OrganisationsInternal
+    public class QualificationsInternal
     {
         private readonly ILogger _logger;
         private readonly IGetOrganisationsSearchUseCase _searchOrganisations;
         private readonly IGetOrganisationByNumberUseCase _getOrganisationByNumber;
 
-        public OrganisationsInternal(ILoggerFactory loggerFactory, IGetOrganisationsSearchUseCase searchOrganisations,
+        public QualificationsInternal(ILoggerFactory loggerFactory, IGetOrganisationsSearchUseCase searchOrganisations,
             IGetOrganisationByNumberUseCase getOrganisationByNumber)
         {
             _logger = loggerFactory.CreateLogger<QualificationsInternal>();
@@ -20,25 +20,25 @@ namespace Ofqual.Common.RegisterAPI.Functions
             _getOrganisationByNumber = getOrganisationByNumber;
         }
 
-        [Function("OrganisationsInternal")]
+        [Function("QualificationsInternal")]
         //Returns the list of qualifications
-        public async Task<HttpResponseData> ListOrganisations([HttpTrigger(AuthorizationLevel.Function, "get")] HttpRequestData req)
+        public async Task<HttpResponseData> ListQualifications([HttpTrigger(AuthorizationLevel.Function, "get")] HttpRequestData req)
         {
             _logger.LogInformation("C# HTTP trigger function processed a request.");
 
             var response = req.CreateResponse(HttpStatusCode.OK);
             response.Headers.Add("Content-Type", "text/plain; charset=utf-8");
 
-            string x = await _getOrganisationByNumber.GetOrganisationByNumber(" ofqual");
+            var x = await _getOrganisationByNumber.GetOrganisationByNumber(" ofqual");
             response.WriteString("Welcome to Azure Functions!" + x);
 
             return response;
         }
 
 
-        [Function("OrganisationInternal")]
+        [Function("QualificationInternal")]
         //Returns a single qualification based on the id parameter provided in the HttpRequestData
-        public async Task<HttpResponseData> GetOrganisation([HttpTrigger(AuthorizationLevel.Function, "get")] HttpRequestData req)
+        public async Task<HttpResponseData> GetQualification([HttpTrigger(AuthorizationLevel.Function, "get")] HttpRequestData req)
         {
 
             _logger.LogInformation("C# HTTP trigger function processed a request.");
@@ -46,11 +46,10 @@ namespace Ofqual.Common.RegisterAPI.Functions
             var response = req.CreateResponse(HttpStatusCode.OK);
             response.Headers.Add("Content-Type", "text/plain; charset=utf-8");
 
-            string x = await _getOrganisationByNumber.GetOrganisationByNumber(" ofqual");
+            var x = await _getOrganisationByNumber.GetOrganisationByNumber(" ofqual");
             response.WriteString("Welcome to Azure Functions!" + x);
 
             return response;
         }
-
     }
 }
