+++ QualificationsRegister/QualificationsRegister.csproj
@@ -12,6 +12,8 @@
     <PackageReference Include="Microsoft.Azure.Functions.Worker.Sdk" Version="1.16.2" />
     <PackageReference Include="Microsoft.ApplicationInsights.WorkerService" Version="2.21.0" />
     <PackageReference Include="Microsoft.Azure.Functions.Worker.ApplicationInsights" Version="1.0.0" />
+    <PackageReference Include="Microsoft.Azure.WebJobs.Extensions" Version="5.0.0" />
+    <PackageReference Include="StackExchange.Redis" Version="2.7.20" />
   </ItemGroup>
   <ItemGroup>
     <None Update="host.json">
